Fix tslint, add more tests to the runner integration:
