+++ package.json
@@ -78,7 +78,7 @@
     "jest": "21",
     "lint-staged": "^4.0.0",
     "madge": "^2.0.0",
-    "prettier": "^1.5.3",
+    "prettier": "^1.7.4",
     "shx": "^0.2.1",
     "ts-jest": "^21",
     "ts-node": "^3.2.1",
